(chore) move remaining NewWorkout functionality to useNewWorkout
